requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Send2Trash>=1.8.0
pywebview>=4.4.0
Pillow>=10.2.0
